requirements.txt
backoff
bokeh
cookiecutter
django
djangorestframework
whitenoise
dj-database-url
django-cors-headers
django-filter
django-guardian
django-import-export
django-model-utils
django-rq
django-shapeshifter
django-tables2
django-taggit
django-test-plus
django-watson
easy-thumbnails
networkx
numpy
pandas
ratelimit
requests
scipy
tqdm
codecov
pytest
pytest-django
